# f1api/openapi/api.yml
openapi: "3.0.0"
info:
  title: F1 API
  version: "1.0.0"
  description: |
    ## Paging and Sorting

    The following query paramters can be used to control the paging and sorting of collections.

    ### `pageSize`

    Defines the number of elements per page.

    **Default value:** 25

    **Notice:** maximum number is 100

    ### `pageNumber`

    Specifies the page number that should be returned.

    **Notice:** first page has the number 0

    **Default value:** 0

    ### `sort`

    Specifies by which properties to sort. <br>
    By default the sort order is ascending. This can be changed by adding `asc` for ascending or `desc` for
    descending separated by a comma after the sort property.

    **Hint:** To sort by more than one property, simply add a new sort parameter to the query.

    ### Example

    GET `/drivers?pageSize=20&pageNumber=0&sort=firstName,asc`

    will return the first page with a maximum of 20 drivers sorted by first name in ascending order.
...
paths:
...
  /seasons:
    post:
      x-internal: true
      summary: Create season
      tags:
        - "Seasons"
        - "SeasonController"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./components/schemas.yml#/Season"
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: "./components/schemas.yml#/Season"
        400:
          $ref: "./components/responses.yml#/400"
        409:
          $ref: "./components/responses.yml#/409"
        422:
          $ref: "./components/responses.yml#/422"
        500:
          $ref: "./components/responses.yml#/500"
    get:
      summary: Get seasons / Get season by seasonYear
      tags:
        - "Seasons"
        - "SeasonController"
      parameters:
        - $ref: "./components/parameters.yml#/seasonYear"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "./components/schemas.yml#/Seasons"
        400:
          $ref: "./components/responses.yml#/400"
        500:
          $ref: "./components/responses.yml#/500"
  /seasons/{seasonId}:
    get:
      summary: Get season
      tags:
        - "Seasons"
        - "SeasonController"
      parameters:
        - $ref: "./components/parameters.yml#/seasonId"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "./components/schemas.yml#/Season"
        400:
          $ref: "./components/responses.yml#/400"
        404:
          $ref: "./components/responses.yml#/404"
        500:
          $ref: "./components/responses.yml#/500"
    patch:
      x-internal: true
      summary: Update season
      tags:
        - "Seasons"
        - "SeasonController"
      parameters:
        - $ref: "./components/parameters.yml#/seasonId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./components/schemas.yml#/Season"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "./components/schemas.yml#/Season"
        400:
          $ref: "./components/responses.yml#/400"
        404:
          $ref: "./components/responses.yml#/404"
        409:
          $ref: "./components/responses.yml#/409"
        422:
          $ref: "./components/responses.yml#/422"
        500:
          $ref: "./components/responses.yml#/500"
  /seasons/{seasonId}/races:
    post:
      x-internal: true
      summary: Create race of a season
      tags:
        - "Races"
        - "SeasonController"
      parameters:
        - $ref: "./components/parameters.yml#/seasonId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./components/schemas.yml#/Race"
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: "./components/schemas.yml#/Race"
        400:
          $ref: "./components/responses.yml#/400"
        404:
          $ref: "./components/responses.yml#/404"
        409:
          $ref: "./components/responses.yml#/409"
        422:
          $ref: "./components/responses.yml#/422"
        500:
          $ref: "./components/responses.yml#/500"
    get:
      summary: Get races of a season / Get race of a season by raceName/date
      tags:
        - "Races"
        - "SeasonController"
      parameters:
        - $ref: "./components/parameters.yml#/seasonId"
        - $ref: "./components/parameters.yml#/raceName"
        - $ref: "./components/parameters.yml#/date"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "./components/schemas.yml#/Races"
        400:
          $ref: "./components/responses.yml#/400"
        404:
          $ref: "./components/responses.yml#/404"
        500:
          $ref: "./components/responses.yml#/500"
  /seasons/{seasonId}/teams:
    get:
      summary: Get teams of a season
      tags:
        - "Teams"
        - "SeasonController"
      parameters:
        - $ref: "./components/parameters.yml#/seasonId"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "./components/schemas.yml#/Teams"
        400:
          $ref: "./components/responses.yml#/400"
        404:
          $ref: "./components/responses.yml#/404"
        500:
          $ref: "./components/responses.yml#/500"
  /seasons/{seasonId}/teams/{teamId}/drivers:
    get:
      summary: Get drivers of a team of a season
      tags:
        - "Drivers"
        - "SeasonController"
      parameters:
        - $ref: "./components/parameters.yml#/seasonId"
        - $ref: "./components/parameters.yml#/teamId"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "./components/schemas.yml#/Drivers"
        400:
          $ref: "./components/responses.yml#/400"
        404:
          $ref: "./components/responses.yml#/404"
        500:
          $ref: "./components/responses.yml#/500"
  /seasons/{seasonId}/teams/{teamId}/engine:
    get:
      summary: Get engine of a team of a season
      tags:
        - "Engines"
        - "SeasonController"
      parameters:
        - $ref: "./components/parameters.yml#/seasonId"
        - $ref: "./components/parameters.yml#/teamId"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "./components/schemas.yml#/Engine"
        400:
          $ref: "./components/responses.yml#/400"
        404:
          $ref: "./components/responses.yml#/404"
        500:
          $ref: "./components/responses.yml#/500"
  /seasons/{seasonId}/drivers:
    get:
      summary: Get drivers of a season
      tags:
        - "Drivers"
        - "SeasonController"
      parameters:
        - $ref: "./components/parameters.yml#/seasonId"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "./components/schemas.yml#/Drivers"
        400:
          $ref: "./components/responses.yml#/400"
        404:
          $ref: "./components/responses.yml#/404"
        500:
          $ref: "./components/responses.yml#/500"
  /seasons/{seasonId}/drivers/{driverId}/teams:
    get:
      summary: Get teams of a driver of a season
      tags:
        - "Teams"
        - "SeasonController"
      parameters:
        - $ref: "./components/parameters.yml#/seasonId"
        - $ref: "./components/parameters.yml#/driverId"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "./components/schemas.yml#/Teams"
        400:
          $ref: "./components/responses.yml#/400"
        404:
          $ref: "./components/responses.yml#/404"
        500:
          $ref: "./components/responses.yml#/500"
  /seasons/{seasonId}/engines:
    get:
      summary: Get engines of a season
      tags:
        - "Engines"
        - "SeasonController"
      parameters:
        - $ref: "./components/parameters.yml#/seasonId"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "./components/schemas.yml#/Engines"
        400:
          $ref: "./components/responses.yml#/400"
        404:
          $ref: "./components/responses.yml#/404"
        500:
          $ref: "./components/responses.yml#/500"
  /seasons/{seasonId}/engines/{engineId}/teams:
    get:
      summary: Get teams which use the engine in the season
      tags:
        - "Teams"
        - "SeasonController"
      parameters:
        - $ref: "./components/parameters.yml#/seasonId"
        - $ref: "./components/parameters.yml#/engineId"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "./components/schemas.yml#/Teams"
        400:
          $ref: "./components/responses.yml#/400"
        404:
          $ref: "./components/responses.yml#/404"
        500:
          $ref: "./components/responses.yml#/500"

  # RaceController
...
  /races/{raceId}:
    get:
      summary: Get race
      tags:
        - "Races"
        - "RaceController"
      parameters:
        - $ref: "./components/parameters.yml#/raceId"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "./components/schemas.yml#/Race"
        400:
          $ref: "./components/responses.yml#/400"
        404:
          $ref: "./components/responses.yml#/404"
        500:
          $ref: "./components/responses.yml#/500"
    patch:
      x-internal: true
      summary: Update race
      tags:
        - "Races"
        - "RaceController"
      parameters:
        - $ref: "./components/parameters.yml#/raceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./components/schemas.yml#/Race"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "./components/schemas.yml#/Race"
        400:
          $ref: "./components/responses.yml#/400"
        404:
          $ref: "./components/responses.yml#/404"
        409:
          $ref: "./components/responses.yml#/409"
        422:
          $ref: "./components/responses.yml#/422"
        500:
          $ref: "./components/responses.yml#/500"
...